Discard backspace keypresses in BeforeInputEventPlugin
Gecko-based browsers may use a `which` value for keypress events on the backspace key.

Discard these keypress events when determining whether the event should be used for `beforeInput` data.

There may be other keys that should be similarly blacklisted, but I don't want to enumerate all of them unless there are actually problems.